Check online status while working with api

+++ src/api/provider.js
@@ -1,3 +1,6 @@
+import nanoid from "nanoid";
+import Point from "../models/point";
+
 export default class Provider {
   constructor(api, backup) {
     this._api = api;
@@ -5,27 +8,56 @@ export default class Provider {
   }
 
   getPoints() {
-    return this._api.getPoints();
+    if (this._isOnLine()) {
+      return this._api.getPoints();
+    }
+
+    return Promise.resolve(Point.parsePoints([]));
   }
 
   getDestinations() {
-    return this._api.getDestinations();
+    if (this._isOnLine()) {
+      return this._api.getDestinations();
+    }
+
+    return Promise.resolve([]);
   }
 
   getOffers() {
-    return this._api.getOffers();
+    if (this._isOnLine()) {
+      return this._api.getOffers();
+    }
+
+    return Promise.resolve([]);
   }
 
   createPoint(point) {
-    return this._api.createPoint(point);
+    if (this._isOnLine()) {
+      return this._api.createPoint(point);
+    }
+
+    const fakeNewPointId = nanoid();
+    const fakeNewPoint = Point.parsePoint(
+        Object.assign({}, point.toRAW(), {id: fakeNewPointId})
+    );
+
+    return Promise.resolve(fakeNewPoint);
   }
 
   updatePoint(id, data) {
-    return this._api.updatePoint(id, data);
+    if (this._isOnLine()) {
+      return this._api.updatePoint(id, data);
+    }
+
+    return Promise.resolve(data);
   }
 
   deletePoint(id) {
-    return this._api.deletePoint(id);
+    if (this._isOnLine()) {
+      return this._api.deletePoint(id);
+    }
+
+    return Promise.resolve();
   }
 
   _isOnLine() {
